fridge_recipes_starter/requirements_mongo.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pymongo>=4.7.0
pandas>=2.2.2
openpyxl>=3.1.2
python-dotenv>=1.0.1
